requirements.txt
python-dotenv
openai
llama-index-core
llama-index-llms-openai
llama-index-readers-wikipedia
llama-index-embeddings-openai
traceloop-sdk
wikipedia
streamlit
nltk
